+++ projects/ngx-mat-lib/src/lib/xmat-models/index.ts
@@ -3,7 +3,6 @@ export * from "./xmat-array-item-from-object.model";
 export * from "./xmat-confirm-dialog.model";
 export * from "./xmat-delayed-http-params.model";
 export * from "./xmat-desc-value.model";
-export * from "./xmat-element-ref.model";
 export * from "./xmat-file.model";
 export * from "./xmat-generic-object";
 export * from "./xmat-http-config.model";