+++ src/app/app.module.ts
@@ -34,7 +34,7 @@ import {
     XmatDialogModule,
     XmatActionTextModule,
     XmatAccordionModule,
-    XmatTimeInputModule,
+    XmatTimeModule,
     XmatSnackBarModule,
     XmatLegendModule,
     // SERVS
@@ -80,7 +80,7 @@ export class XmatCdkMatImportsModule {
         XmatAccordionModule,
         XmatDialogModule,
         XmatLegendModule,
-        XmatTimeInputModule,
+        XmatTimeModule,
         XmatSnackBarModule,
     ],
     providers: [